refactored sort by to be on all list pages, not as SecondaryNav
