+++ src/client/features/students/StudentsByLastName.jsx
@@ -15,10 +15,33 @@ export default function Students() {
   const [filter, setFilter] = useState("");
   const searchRegex = new RegExp(filter, "i");
 
+  const handleSort = (e) => {
+    e.preventDefault();
+    console.log(e.target.value);
+    if (e.target.value === "last-name") {
+      navigate("/students/sortedbylastname");
+    }
+    if (e.target.value === "first-name") {
+      navigate("/students");
+    }
+    if (e.target.value === "gpa") {
+      navigate("/students/sortedbygpa");
+    }
+  };
+
   return isLoading ? (
     <h2>Loading students...</h2>
   ) : (
     <main>
+      <div>
+        <label>Sort By</label>
+        <select id="sort-by" onChange={handleSort}>
+          <option value="Please select"></option>
+          <option value="first-name">First Name</option>
+          <option value="last-name">Last Name</option>
+          <option value="gpa">GPA</option>
+        </select>
+      </div>
       <br />
       <form>
         <input
